Refactor Base Exceptions
