+++ core/dbt/adapters/exceptions/__init__.py
@@ -1,0 +1,2 @@
+from dbt.adapters.exceptions.compilation import *  # noqa
+from dbt.adapters.exceptions.alias import *  # noqa
