+++ core/dbt/adapters/factory.py
@@ -10,7 +10,7 @@ from dbt.adapters.protocol import AdapterConfig, AdapterProtocol, RelationProtoc
 from dbt.adapters.contracts.connection import AdapterRequiredConfig, Credentials
 from dbt.common.events.functions import fire_event
 from dbt.common.events.types import AdapterImportError, PluginLoadError, AdapterRegistered
-from dbt.exceptions import DbtInternalError, DbtRuntimeError
+from dbt.common.exceptions import DbtInternalError, DbtRuntimeError
 from dbt.include.global_project import PACKAGE_PATH as GLOBAL_PROJECT_PATH
 from dbt.include.global_project import PROJECT_NAME as GLOBAL_PROJECT_NAME
 from dbt.semver import VersionSpecifier
