+++ tests/adapter/dbt/tests/adapter/utils/test_validate_sql.py
@@ -41,7 +41,7 @@ class BaseValidateSqlMethod:
     def expected_exception(self) -> Type[Exception]:
         """Returns the Exception type thrown by a failed query.
 
-        Defaults to dbt.exceptions.DbtRuntimeError because that is the most common
+        Defaults to dbt.common.exceptions.DbtRuntimeError because that is the most common
         base exception for adapters to throw."""
         return DbtRuntimeError
