+++ tests/functional/dependencies/test_local_dependency.py
@@ -197,7 +197,7 @@ class TestMissingDependency(object):
 
     def test_missing_dependency(self, project):
         # dbt should raise a runtime exception
-        with pytest.raises(dbt.exceptions.DbtRuntimeError):
+        with pytest.raises(dbt.common.exceptions.DbtRuntimeError):
             run_dbt(["compile"])
 
 
@@ -352,7 +352,7 @@ class TestSimpleDependencyDuplicateName(BaseDependencyTest):
 
     def test_local_dependency_same_name_sneaky(self, prepare_dependencies, project):
         shutil.copytree("duplicate_dependency", "./dbt_packages/duplicate_dependency")
-        with pytest.raises(dbt.exceptions.CompilationError):
+        with pytest.raises(dbt.common.exceptions.CompilationError):
             run_dbt(["compile"])
 
         # needed to avoid compilation errors from duplicate package names in test autocleanup
