+++ tests/functional/docs/test_duplicate_docs_block.py
@@ -31,5 +31,5 @@ class TestDuplicateDocsBlock:
         }
 
     def test_duplicate_doc_ref(self, project):
-        with pytest.raises(dbt.exceptions.CompilationError):
+        with pytest.raises(dbt.common.exceptions.CompilationError):
             run_dbt(expect_pass=False)
