+++ tests/functional/docs/test_invalid_doc_ref.py
@@ -43,5 +43,5 @@ class TestInvalidDocRef:
 
     def test_invalid_doc_ref(self, project):
         # The run should fail since we could not find the docs reference.
-        with pytest.raises(dbt.exceptions.CompilationError):
+        with pytest.raises(dbt.common.exceptions.CompilationError):
             run_dbt(expect_pass=False)
