+++ tests/functional/macros/test_macros.py
@@ -118,7 +118,7 @@ class TestAdapterMacroNoDestination:
         return {"my_macros.sql": macros__no_default_macros}
 
     def test_invalid_macro(self, project):
-        with pytest.raises(dbt.exceptions.CompilationError) as exc:
+        with pytest.raises(dbt.common.exceptions.CompilationError) as exc:
             run_dbt()
 
         assert "In dispatch: No macro named 'dispatch_to_nowhere' found" in str(exc.value)
@@ -255,7 +255,7 @@ class TestMisnamedMacroNamespace:
     ):
         run_dbt(["deps"])
 
-        with pytest.raises(dbt.exceptions.CompilationError) as exc:
+        with pytest.raises(dbt.common.exceptions.CompilationError) as exc:
             run_dbt()
 
         assert "In dispatch: No macro named 'cowsay' found" in str(exc.value)
@@ -271,7 +271,7 @@ class TestAdapterMacroDeprecated:
         return {"macro.sql": macros__deprecated_adapter_macro}
 
     def test_invalid_macro(self, project):
-        with pytest.raises(dbt.exceptions.CompilationError) as exc:
+        with pytest.raises(dbt.common.exceptions.CompilationError) as exc:
             run_dbt()
 
         assert 'The "adapter_macro" macro has been deprecated' in str(exc.value)
